+++ runtime/src/tests.rs
@@ -7,23 +7,11 @@ use frame_support::{assert_noop, assert_ok};
 use frame_system::EventRecord;
 use frame_system::RawOrigin;
 use pallet_scheduler::Event as SchedulerEvent;
-use runtime_utils::{ExtBuilder, RollTo, Trivial};
-use sp_core::Pair;
-use sp_runtime::{
-	traits::{BadOrigin, IdentifyAccount},
-	AccountId32, MultiSigner,
-};
+use runtime_utils::{generate_account, ExtBuilder, RollTo, Trivial};
+use sp_runtime::traits::BadOrigin;
 use std::default::Default;
 
 mod staking;
-
-fn generate_account(seed: &str) -> AccountId32 {
-	let seed = seed.bytes().cycle().take(32).collect::<Vec<_>>();
-	let key_pair = sp_core::ecdsa::Pair::from_seed_slice(seed.as_slice()).unwrap();
-	let pkey = key_pair.public();
-	let signer: MultiSigner = pkey.into();
-	signer.into_account()
-}
 
 #[test]
 fn pallet_scheduler_works() {
